+++ GameLib.Plugin/GameLib.Plugin.Rockstar/RockstarLauncher.cs
@@ -1,0 +1,89 @@
+﻿using Gamelib.Core.Util;
+using GameLib.Core;
+using Microsoft.Win32;
+using System.ComponentModel.Composition;
+using System.Drawing;
+using System.Runtime.InteropServices;
+
+namespace GameLib.Plugin.Rockstar;
+
+[Guid("7A24C50D-4D48-4B56-BEF4-5EBDC0B107EC")]
+[Export(typeof(ILauncher))]
+public class RockstarLauncher : ILauncher
+{
+    [ImportingConstructor]
+    public RockstarLauncher(LauncherOptions? launcherOptions)
+    {
+        LauncherOptions = launcherOptions ?? new LauncherOptions();
+    }
+
+    #region Interface implementations
+    public LauncherOptions LauncherOptions { get; }
+
+    public Guid Id => GetType().GUID;
+
+    public string Name => "Rockstar Games";
+
+    public Image Logo => Properties.Resources.Logo128px;
+
+    public bool IsInstalled { get; private set; }
+
+    public bool IsRunning => ProcessUtil.IsProcessRunning(ExecutablePath);
+
+    public string InstallDir { get; private set; } = string.Empty;
+
+    public string ExecutablePath { get; private set; } = string.Empty;
+
+    public string Executable { get; private set; } = string.Empty;
+
+    public Icon? ExecutableIcon => PathUtil.GetFileIcon(ExecutablePath);
+
+    public IEnumerable<IGame> Games { get; private set; } = Enumerable.Empty<IGame>();
+
+    public bool Start() => IsRunning || ProcessUtil.StartProcess(ExecutablePath);
+
+    public void Stop() => ProcessUtil.StopProcess(ExecutablePath);
+
+    public void Refresh(CancellationToken cancellationToken = default)
+    {
+        ExecutablePath = string.Empty;
+        Executable = string.Empty;
+        InstallDir = string.Empty;
+        IsInstalled = false;
+        Games = Enumerable.Empty<IGame>();
+
+        ExecutablePath = GetExecutable() ?? string.Empty;
+        if (!string.IsNullOrEmpty(ExecutablePath))
+        {
+            Executable = Path.GetFileName(ExecutablePath);
+            InstallDir = Path.GetDirectoryName(ExecutablePath) ?? string.Empty;
+            IsInstalled = File.Exists(ExecutablePath);
+            Games = RockstarGameFactory.GetGames(this, cancellationToken);
+        }
+    }
+    #endregion
+
+    #region Private methods
+    private static string? GetExecutable()
+    {
+        string? executablePath = null;
+
+        executablePath ??= RegistryUtil.GetShellCommand("rockstar");
+        executablePath ??= RegistryUtil.GetValue(RegistryHive.LocalMachine, @"Software\Rockstar Games\Launcher", "InstallFolder");
+
+        executablePath = PathUtil.Sanitize(executablePath);
+
+        if (!string.IsNullOrEmpty(executablePath) && !PathUtil.IsExecutable(executablePath))
+        {
+            executablePath = Path.Combine(executablePath, "Launcher.exe");
+        }
+
+        if (!PathUtil.IsExecutable(executablePath))
+        {
+            executablePath = null;
+        }
+
+        return executablePath;
+    }
+    #endregion
+}